pyproject.toml
[project]
name = "testtaskfast"
version = "0.1.0"
description = "Calendar NLP Assistant with Multilingual Support"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.2",
    "fastapi[all]>=0.127.0",
    "passlib[argon2-cffi]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.45",
    "websockets>=12.0",
    "pydantic>=2.0.0",
    # NLP dependencies
    #"torch>=2.0.0",
    "transformers>=4.30.0",
    "dateparser>=1.2.0",
    "python-dateutil>=2.8.2",
    "pytz>=2023.3",
    #"rrule==0.0.1",
    "iso8601>=2.1.0",
    "scikit-learn>=1.3.0",
    "numpy>=1.24.0",
    # Additional
    "redis>=5.0.0",
    "celery>=5.3.0",
    "apscheduler>=3.10.0",
    "aio-pika>=9.5.5",
    "aiormq>=6.8.1"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "httpx>=0.25.0",
]
test = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
]